pyproject.toml
[project]
name = "llm-arena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "dotenv>=0.9.9",
    "openai>=1.102.0",
    "pillow>=11.3.0",
    "zai-sdk>=0.0.3.3",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
